Update generated library support for org.apache.ivy:ivy:2.5.2
