+++ metadata/org.apache.ivy/ivy/2.5.2/reachability-metadata.json
@@ -1,0 +1,85 @@
+{
+  "reflection" : [
+    {
+      "condition" : {
+        "typeReached" : "org.apache.ivy.util.url.IvyAuthenticator"
+      },
+      "type" : "java.net.Authenticator",
+      "fields" : [
+        {
+          "name" : "theAuthenticator"
+        }
+      ],
+      "methods" : [
+        {
+          "name" : "getDefault",
+          "parameterTypes" : []
+        }
+      ]
+    },
+    {
+      "condition" : {
+        "typeReached" : "org.apache.ivy.core.settings.IvySettings"
+      },
+      "type" : "org.apache.ivy.core.settings.IvySettings",
+      "allPublicMethods" : true
+    },
+    {
+      "condition" : {
+        "typeReached" : "org.apache.ivy.core.settings.IvySettings"
+      },
+      "type" : "org.apache.ivy.core.settings.IvyPattern",
+      "allPublicConstructors" : true,
+      "allPublicMethods" : true
+    },
+    {
+      "condition" : {
+        "typeReached" : "org.apache.ivy.core.settings.IvySettings"
+      },
+      "type" : "org.apache.ivy.plugins.matcher.GlobPatternMatcher",
+      "fields" : [
+        {
+          "name" : "INSTANCE"
+        }
+      ]
+    },
+    {
+      "condition" : {
+        "typeReached" : "org.apache.ivy.core.settings.IvySettings"
+      },
+      "type" : "org.apache.ivy.plugins.resolver.IBiblioResolver",
+      "allPublicConstructors" : true,
+      "allPublicMethods" : true
+    },
+    {
+      "condition" : {
+        "typeReached" : "org.apache.ivy.core.settings.IvySettings"
+      },
+      "type" : "org.apache.ivy.plugins.resolver.FileSystemResolver",
+      "allPublicConstructors" : true,
+      "allPublicMethods" : true
+    },
+    {
+      "condition" : {
+        "typeReached" : "org.apache.ivy.core.settings.IvySettings"
+      },
+      "type" : "org.apache.ivy.plugins.resolver.ChainResolver",
+      "allPublicConstructors" : true,
+      "allPublicMethods" : true
+    }
+  ],
+  "resources" : [
+    {
+      "condition" : {
+        "typeReached" : "org.apache.ivy.core.settings.IvySettings"
+      },
+      "glob" : "org/apache/ivy/core/settings/*"
+    },
+    {
+      "condition" : {
+        "typeReached" : "org.apache.ivy.osgi.core.ExecutionEnvironmentProfileProvider"
+      },
+      "glob" : "org/apache/ivy/osgi/core/jvm-packages.properties"
+    }
+  ]
+}
