+++ stats/org.apache.ivy/ivy/2.5.2/stats.json
@@ -1,0 +1,42 @@
+{
+  "versions" : [ {
+    "version" : "2.5.2",
+    "dynamicAccess" : {
+      "breakdown" : {
+        "reflection" : {
+          "coverageRatio" : 0.928571,
+          "coveredCalls" : 26,
+          "totalCalls" : 28
+        },
+        "resources" : {
+          "coverageRatio" : 1.0,
+          "coveredCalls" : 3,
+          "totalCalls" : 3
+        }
+      },
+      "coverageRatio" : 0.935484,
+      "coveredCalls" : 29,
+      "totalCalls" : 31
+    },
+    "libraryCoverage" : {
+      "instruction" : {
+        "covered" : 16950,
+        "missed" : 114825,
+        "ratio" : 0.128628,
+        "total" : 131775
+      },
+      "line" : {
+        "covered" : 3874,
+        "missed" : 24602,
+        "ratio" : 0.136044,
+        "total" : 28476
+      },
+      "method" : {
+        "covered" : 927,
+        "missed" : 4899,
+        "ratio" : 0.159114,
+        "total" : 5826
+      }
+    }
+  } ]
+}
