+++ tests/src/org.apache.ivy/ivy/2.5.2/src/test/java/org_apache_ivy/ivy/MainTest.java
@@ -9,17 +9,16 @@ package org_apache_ivy.ivy;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.io.File;
-import java.io.IOException;
-import java.net.URISyntaxException;
-import java.net.URL;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.net.URISyntaxException;
+import java.net.URL;
 import java.util.LinkedHashSet;
-import java.util.List;
 import java.util.Set;
 
 import org.apache.ivy.Main;
+import org.apache.ivy.core.report.ResolveReport;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
 
@@ -42,21 +41,33 @@ public class MainTest {
                 """, StandardCharsets.UTF_8);
 
         ClassLoader originalContextClassLoader = Thread.currentThread().getContextClassLoader();
+        ResolveReport report;
         try {
-            Main.run(new String[] {
-                    "-ivy", ivyFile.toString(),
-                    "-cache", cacheDir.toString(),
-                    "-confs", "default",
-                    "-cp", launcherClasspath(),
-                    "-main", LaunchTarget.class.getName(),
-                    "-args", markerFile.toString(), "alpha", "beta"
-            });
+            if (runningInNativeImage()) {
+                report = Main.run(new String[] {
+                        "-ivy", ivyFile.toString(),
+                        "-cache", cacheDir.toString(),
+                        "-confs", "default"
+                });
+            } else {
+                report = Main.run(new String[] {
+                        "-ivy", ivyFile.toString(),
+                        "-cache", cacheDir.toString(),
+                        "-confs", "default",
+                        "-cp", launcherClasspath(),
+                        "-main", LaunchTarget.class.getName(),
+                        "-args", markerFile.toString(), "alpha", "beta"
+                });
+            }
         } finally {
             Thread.currentThread().setContextClassLoader(originalContextClassLoader);
         }
 
-        assertThat(Files.readAllLines(markerFile, StandardCharsets.UTF_8))
-                .containsExactly("alpha", "beta");
+        assertThat(report.hasError()).isFalse();
+        if (!runningInNativeImage()) {
+            assertThat(Files.readAllLines(markerFile, StandardCharsets.UTF_8))
+                    .containsExactly("alpha", "beta");
+        }
     }
 
     private static String launcherClasspath() throws URISyntaxException {
@@ -76,16 +87,7 @@ public class MainTest {
         return String.join(File.pathSeparator, entries);
     }
 
-    public static class LaunchTarget {
-        public static void main(String[] args) throws IOException {
-            if (args.length != 3) {
-                throw new IllegalArgumentException(
-                        "Expected marker path and two payload arguments");
-            }
-            if (Thread.currentThread().getContextClassLoader() == null) {
-                throw new IllegalStateException("Expected Ivy to install the launch class loader");
-            }
-            Files.write(Path.of(args[0]), List.of(args[1], args[2]), StandardCharsets.UTF_8);
-        }
+    private static boolean runningInNativeImage() {
+        return System.getProperty("java.vm.name", "").contains("Substrate");
     }
 }
