+++ tests/src/org.apache.ivy/ivy/2.5.2/src/test/resources/META-INF/native-image/ivy-tests-only/reflect-config.json
@@ -1,19 +1,5 @@
 [
   {
-    "name": "java.net.Authenticator",
-    "fields": [
-      {
-        "name": "theAuthenticator"
-      }
-    ],
-    "methods": [
-      {
-        "name": "getDefault",
-        "parameterTypes": []
-      }
-    ]
-  },
-  {
     "name": "org_apache_ivy.ivy.MemoryUtilTest$MeasuredObject",
     "allDeclaredConstructors": true
   }
