+++ tests/src/org.apache.ivy/ivy/2.5.2/user-code-filter.json
@@ -1,10 +1,10 @@
 {
-  "rules": [
+  "rules" : [
     {
-      "excludeClasses": "**"
+      "excludeClasses" : "**"
     },
     {
-      "includeClasses": "org.apache.ivy.**"
+      "includeClasses" : "org.apache.ivy.**"
     }
   ]
 }
